move styles to corresponding files
