+++ src/components/DetailsDrillDownAlert/DetailsAlertsMetrics.jsx
@@ -22,7 +22,7 @@ import PropTypes from 'prop-types'
 import { useDispatch, useSelector } from 'react-redux'
 import { isEmpty } from 'lodash'
 
-import ApplicationMetricCard from '../DetailsMetrics/MetricsCards/ApplicationMetricCard'
+import ApplicationMetricCard from '../DetailsMetrics/MetricsCards/ApplicationMetricCard/ApplicationMetricCard'
 import DatePicker from '../../common/DatePicker/DatePicker'
 import NoData from '../../common/NoData/NoData'
 import NoMetricData from '../DetailsMetrics/MetricsCards/NoMetricData/NoMetricData'
