+++ src/components/DetailsMetrics/MetricsCards/InvocationsMetricCard/InvocationsMetricCard.scss
@@ -1,0 +1,201 @@
+@use 'igz-controls/scss/colors';
+@use 'igz-controls/scss/borders';
+
+.metrics__card-invocation {
+  position: relative;
+  z-index: 1;
+  width: 100%;
+
+  .metrics__card-body {
+    aspect-ratio: 5;
+
+    &-invocation {
+      flex: 1;
+      border: 1px solid transparent;
+
+      canvas {
+        width: 100% !important;
+        height: 100% !important;
+      }
+
+      canvas.hidden-canvas {
+        position: absolute;
+        top: 0;
+        left: 0;
+        opacity: 0;
+      }
+    }
+  }
+
+  &__toggle-icon {
+    position: absolute;
+    right: 5px;
+    bottom: 5px;
+    align-self: flex-start;
+    visibility: hidden;
+    opacity: 0;
+    transition: all 0.2s ease-in-out;
+  }
+
+  &:hover {
+    .metrics__card-invocation__toggle-icon {
+      visibility: visible;
+      opacity: 1;
+
+      .round-icon-cp__circle {
+        &::before {
+          opacity: 0.4;
+        }
+
+        &:hover::before {
+          opacity: 1;
+        }
+      }
+    }
+  }
+
+  &-header {
+    display: flex;
+    flex: 1;
+    align-items: baseline;
+    justify-content: end;
+    height: 40px;
+    color: colors.$primary;
+    font-weight: 500;
+    font-size: 14px;
+
+    &__drift-icon-container {
+      align-self: stretch;
+      margin-right: 2px;
+
+      svg {
+        box-sizing: content-box;
+        width: 17px;
+        height: 17px;
+        padding-top: 9px;
+      }
+    }
+
+    &__drift_up {
+      color: colors.$brightTurquoise;
+    }
+
+    &__drift_down {
+      color: colors.$ceriseRed;
+    }
+
+    &__selected-date {
+      margin-left: 4px;
+      color: colors.$topaz;
+    }
+
+    &__total-title {
+      margin: 0 6px 0 12px;
+      font-weight: 700;
+    }
+
+    &__total-score {
+      font-weight: 700;
+      font-size: 24px;
+    }
+  }
+
+  &-content {
+    flex: 0 0 0;
+    flex-wrap: wrap;
+    align-content: center;
+    align-items: baseline;
+    justify-content: center;
+    width: 0;
+    font-weight: 700;
+    font-size: 18px;
+    visibility: hidden;
+    opacity: 0;
+    transition: flex 0.2s;
+
+    &-visible {
+      display: flex;
+      flex: 0 1 auto;
+      width: fit-content;
+      margin-right: 15px;
+      visibility: visible;
+      opacity: 1;
+      transition:
+        flex 0.2s,
+        visibility 0s,
+        opacity 1s;
+    }
+
+    &-title {
+      margin-right: 24px;
+      color: colors.$primary;
+      font-size: 18px;
+      text-align: center;
+    }
+
+    &-container {
+      display: flex;
+      gap: 2px;
+      align-items: baseline;
+      font-weight: 700;
+      font-size: 14px;
+
+      &__drift-icon {
+        align-self: stretch;
+
+        svg {
+          box-sizing: content-box;
+          width: 17px;
+          height: 17px;
+          margin-top: -1px;
+        }
+      }
+
+      &__drift_up {
+        margin-right: 4px;
+        color: colors.$brightTurquoise;
+      }
+
+      &__drift_down {
+        margin-right: 4px;
+        color: colors.$ceriseRed;
+      }
+    }
+
+    &-data {
+      display: flex;
+      align-items: baseline;
+      margin-left: 15px;
+
+      &__total-title {
+        margin-right: 7px;
+        font-weight: 700;
+        font-size: 14px;
+      }
+
+      &__total-score {
+        font-weight: 700;
+        font-size: 24px;
+      }
+    }
+  }
+
+  .metrics__card {
+    padding-top: 12px;
+    padding-bottom: 0;
+  }
+
+  &_collapsed .stats-card__row {
+    height: 0;
+    min-height: 0 !important;
+    margin-bottom: 0 !important;
+    opacity: 0;
+    transition: all 0.2s ease-in-out;
+  }
+
+  &_expanded .stats-card__row {
+    height: 40px;
+    opacity: 1;
+    transition: all 0.2s ease-in-out;
+  }
+}
