+++ src/components/DetailsMetrics/MetricsCards/MetricsCards.scss
@@ -1,0 +1,144 @@
+@use 'igz-controls/scss/colors';
+@use 'igz-controls/scss/borders';
+
+.metrics__card {
+  width: 100%;
+  padding-bottom: 35px;
+  border: borders.$primaryBorder;
+  box-shadow: none;
+
+  &.metrics__card-metric {
+    .metrics__card-body {
+      min-height: 130px;
+      aspect-ratio: 7;
+    }
+  }
+
+  .metrics__card-row {
+    display: flex;
+    flex: 1 0 100%;
+  }
+
+  &-drift-status {
+    display: inline-block;
+    width: 10px;
+    height: 10px;
+    margin-left: 8px;
+    border-radius: 50%;
+  }
+
+  &-drift-status-possible_drift {
+    background-color: colors.$grandis;
+  }
+
+  &-drift-status-drift_detected {
+    background-color: colors.$ceriseRed;
+  }
+
+  &-drift-status-no_drift {
+    background-color: colors.$brightTurquoise;
+  }
+
+  &-drift-no_status {
+    background-color: colors.$doveGray;
+  }
+
+  &-header {
+    position: relative;
+    top: -5px;
+    display: flex;
+    flex-direction: row;
+    gap: 8px;
+    align-items: center;
+    justify-content: space-between;
+    padding-left: 5px;
+    font-weight: 500;
+    transition: all 0.2s linear;
+
+    &-result-kind {
+      display: flex;
+      gap: 4px;
+      align-self: center;
+      justify-content: center;
+      width: max-content;
+      overflow: hidden;
+      color: colors.$topaz;
+      font-weight: 500;
+      font-size: 14px;
+      font-family: Roboto, sans-serif;
+      font-style: normal;
+      line-height: normal;
+      text-overflow: ellipsis;
+    }
+
+    &-data {
+      color: colors.$topaz;
+      font-weight: 700;
+      font-size: 18px;
+      text-align: center;
+      leading-trim: both;
+      text-edge: cap;
+    }
+
+    &-label {
+      font-size: 14px;
+    }
+  }
+
+  &-body {
+    position: relative;
+    display: flex;
+    flex: 1;
+    align-self: stretch;
+    box-sizing: border-box;
+    width: 100%;
+
+    &-collapsed {
+      height: 80px;
+    }
+
+    .chart-item {
+      flex: 1;
+
+      canvas {
+        width: 100%;
+        height: 100%;
+      }
+    }
+
+    &-bar {
+      flex: 0 1 25%;
+      min-width: 210px;
+      margin-right: 15px;
+
+      canvas {
+        width: 100%;
+        height: 100%;
+      }
+
+      canvas.hidden-canvas {
+        position: absolute;
+        top: 0;
+        left: 0;
+        opacity: 0;
+      }
+    }
+
+    &-line {
+      flex: 1 1 0;
+      margin-top: 3px;
+
+      canvas {
+        width: 100% !important;
+        height: 100% !important;
+      }
+
+      canvas.hidden-canvas {
+        position: absolute;
+        top: 0;
+        left: 0;
+        opacity: 0;
+      }
+    }
+  }
+}
